==> cardiac_eqtls/4.2.differential_expression_analysis.ipynb <==
setwd("/frazer01/projects/CARDIPS/analysis/cardiac_eqtls")

source("script/packages.R"  )
source("script/input_data.R")
source("script/functions.R" )


# Analysis of differential expression results

gene_info               = fread("pipeline/1.2.expression/gene_info.txt"   , sep = "\t", header = TRUE, data.table = FALSE)
isof_info               = fread("pipeline/1.2.expression//isoform_info.txt", sep = "\t", header = TRUE, data.table = FALSE)
gene_info$transcript_id = gene_info$gene_id

read_diffexp = function(name, tissue1, tissue2, gene_info)
{
    indata               = add_rownames(fread(paste("pipeline/4.1.differential_expression", paste("diffexp", name, tissue1, tissue2, "txt", sep = "."), sep = "/"), sep = "\t", header = TRUE, data.table = FALSE))
    indata$transcript_id = rownames(indata)
    indata               = indata[order(indata$pval),]
    indata               = merge(gene_info[,c("transcript_id", "gene_id", "gene_name", "gene_type")], indata)
    indata$qval          = p.adjust(indata$pval, method = "bonferroni")
    indata$tissue1       = tissue1
    indata$tissue2       = tissue2
    indata$type          = name
    indata$diffexp       = FALSE
    
    indata[indata$qval < 0.05, "diffexp"] = TRUE
    
    message(paste(name, tissue1, tissue2, nrow(indata), nrow(indata[indata$qval < 0.05,])))
    return(indata)
}

diffexp = as.data.frame(rbindlist(lapply(c("ipsc_cvpc", "heart"), function(tissue1)
{
	out = as.data.frame(rbindlist(lapply(c("heart", "arteria"), function(tissue2)
	{
		if(tissue1 != tissue2)
		{
			indata_gene    = read_diffexp("gene_tpm"   , tissue1, tissue2, gene_info)
			indata_iso_tpm = read_diffexp("isoform_tpm", tissue1, tissue2, isof_info)
			indata_iso_use = read_diffexp("isoform_use", tissue1, tissue2, isof_info)
            
            return(rbind(indata_gene, indata_iso_tpm, indata_iso_use))
		}
	})), stringsAsFactors = FALSE)
    
     return(out)
})), stringsAsFactors = FALSE)

#name    = "gene_tpm"
#tissue1 = "ipsc_cvpc"
#tissue2 = "heart"
#head(read_diffexp("gene_tpm"   , tissue1, tissue2, gene_info))
#head(read_diffexp("isoform_use", tissue1, tissue2))

fwrite(diffexp, "pipeline/4.1.differential_expression/diffexp.txt", sep = "\t", col.names = TRUE, row.names = FALSE)

a = diffexp[diffexp$tissue1 == "ipsc_cvpc" & diffexp$tissue2 == "heart" & diffexp$type == "isoform_tpm" & diffexp$gene_name %in% c("SCN5A", "TNNT2", "TTN", "ABLIM1"),]
a[order(a$pval),]

str(diffexp)

# Add filters for expression levels

metadata              =              fread("pipeline/3.1.covariates/metadata.txt"              , sep = "\t", header = TRUE , data.table = FALSE)
covariates            = add_rownames(fread("pipeline/3.1.covariates/covariates.txt"            , sep = "\t", header = TRUE , data.table = FALSE))
gene_tpm_expressed    = add_rownames(fread("pipeline/1.2.expression/tpm_gene.expressed.txt"    , sep = "\t", header = TRUE , data.table = FALSE))
isof_tpm_expressed    = add_rownames(fread("pipeline/1.2.expression/tpm_isoform.expressed.txt" , sep = "\t", header = TRUE , data.table = FALSE))
isof_use_expressed    = add_rownames(fread("pipeline/1.2.expression/use_isoform.expressed.txt" , sep = "\t", header = TRUE , data.table = FALSE))
gene_tpm_normalized   = add_rownames(fread("pipeline/1.2.expression/tpm_gene.normalized.txt"   , sep = "\t", header = TRUE , data.table = FALSE))
isof_tpm_normalized   = add_rownames(fread("pipeline/1.2.expression/tpm_isoform.normalized.txt", sep = "\t", header = TRUE , data.table = FALSE))
isof_use_normalized   = add_rownames(fread("pipeline/1.2.expression/use_isoform.normalized.txt", sep = "\t", header = TRUE , data.table = FALSE))


add_diffexp = function(diffexp, name, tissue1, tissue2, gene_info, covariates, tpm)
{
    indata               = diffexp[diffexp$type == name & diffexp$tissue1 == tissue1 & diffexp$tissue2 == tissue2,]
    tpm1                 = tpm[indata$transcript_id,rownames(covariates[covariates[,tissue1] == 1,])]
    tpm2                 = tpm[indata$transcript_id,rownames(covariates[covariates[,tissue2] == 1,])]
    expdata              = data.frame(transcript_id = indata$transcript_id, tissue1_exp = rowMeans(tpm1[indata$transcript_id,]), tissue2_exp = rowMeans(tpm2[indata$transcript_id,]))
    expdata$delta        = expdata$tissue1_exp - expdata$tissue2_exp
    
    #if(name == "isoform_use")
    #{
        expdata$log2r        = log2(expdata$tissue1_exp / expdata$tissue2_exp)
        expdata$log2r_filter = FALSE

        expdata[is.na(expdata$log2r) == FALSE & abs(expdata$log2r) > 2, "log2r_filter"] = TRUE
    #}else
    #{
    #    expdata$log2r        = 0
    #    expdata$log2r_filter = FALSE
    #}
    
    outdata = merge(indata, expdata)
   
    message(paste(name, tissue1, tissue2, nrow(outdata), nrow(outdata[outdata$diffexp == TRUE,]), nrow(outdata[outdata$diffexp == TRUE & outdata$log2r_filter == TRUE,])))
    return(outdata)
}

diffexp_add = as.data.frame(rbindlist(lapply(c("ipsc_cvpc", "heart"), function(tissue1)
{
	out = as.data.frame(rbindlist(lapply(c("heart", "arteria"), function(tissue2)
	{
		if(tissue1 != tissue2)
		{
			indata_gene    = add_diffexp(diffexp, "gene_tpm"   , tissue1, tissue2, gene_info, covariates, gene_tpm_expressed)
			indata_iso_tpm = add_diffexp(diffexp, "isoform_tpm", tissue1, tissue2, isof_info, covariates, isof_tpm_expressed)
			indata_iso_use = add_diffexp(diffexp, "isoform_use", tissue1, tissue2, isof_info, covariates, isof_use_expressed)
            
            return(rbind(indata_gene, indata_iso_tpm, indata_iso_use))
		}
	})), stringsAsFactors = FALSE)
    
     return(out)
})), stringsAsFactors = FALSE)

#name    = "gene_tpm"
#tissue1 = "heart"
#tissue2 = "arteria"
#
#head(read_diffexp("gene_tpm"   , tissue1, tissue2, gene_info))

fwrite(diffexp_add, "pipeline/4.1.differential_expression/diffexp.txt", sep = "\t", col.names = TRUE, row.names = FALSE)

str(diffexp_add)
