==> cardiac_eqtls/4.20.test_bydomain.ipynb <==
setwd("/frazer01/projects/CARDIPS/analysis/cardiac_eqtls")

source("script/packages.R"  )
source("script/input_data.R")
source("script/functions.R" )
source("script/colors.R"    )


dir.create("pipeline/4.10.diffexp_eclip_enrichment"    , showWarnings = FALSE)
dir.create("pipeline/4.10.diffexp_eclip_enrichment/tmp", showWarnings = FALSE)


geneinfo_genes    = fread("pipeline/1.2.expression/gene_info.txt"   , sep = "\t", header = TRUE , data.table = FALSE)
geneinfo_isoforms = fread("pipeline/1.2.expression/isoform_info.txt", sep = "\t", header = TRUE , data.table = FALSE)


diffgene = fread(paste("pipeline/4.1.differential_expression", "diffexp.txt"     , sep = "/"), sep = "\t", header = TRUE, data.table = FALSE)
diffcell = fread(paste("pipeline/4.1.differential_expression", "diffexp_cell.txt", sep = "/"), sep = "\t", header = TRUE, data.table = FALSE)

diffgene_gene = diffgene[diffgene$type == "gene_tpm",]

diffgene = diffgene[diffgene$type == "isoform_use",]
diffcell = diffcell[diffcell$type == "isoform_use",]




domains      = fread("/reference/public/Prot2HG//prot2hg.txt"                        , sep = ";" , header = TRUE, data.table = FALSE)
exons_unique = fread("pipeline/4.10.diffexp_eclip_enrichment/exons_unique.txt"       , sep = "\t", header = TRUE, data.table = FALSE)
exons2domain = fread("pipeline/4.10.diffexp_eclip_enrichment/exons_unique2domain.txt", sep = "\t", header = TRUE, data.table = FALSE)


exons2domain_single = as.data.frame(rbindlist(lapply(1:nrow(exons_unique), function(ii)
{
    gene_id = unlist(strsplit(exons_unique[ii, "gene_id"], "\\."))[[1]]
    from    = exons_unique[ii, "start"  ]
    to      = exons_unique[ii, "end"    ]
    domain  = sort(unique(domains[domains$ensembl == gene_id & ((domains$chr_start >= from & domains$chr_start <= to) | (domains$chr_end >= from & domains$chr_end <= to)), "feature_name"]))
    
    if(length(domain) > 0)
    {
        out        = exons_unique[rep(ii, length(domain)),]
        out$domain = domain
    }else
    {
        out        = exons_unique[ii,]
        out$domain = ""
    }
    
    return(out)
})), stringsAsFactors = FALSE)

fwrite(exons2domain_single, "pipeline/4.10.diffexp_eclip_enrichment/exons_unique2domain_single.txt", sep = "\t", col.names = TRUE, row.names = FALSE)



domains2test = sort(unique(exons2domain_single[ exons2domain_single$domain != "", "domain"]))
exons        = unique(exons2domain_single[, c("gene_id", "tissue", "tissue1", "tissue2")])

calculate_enrichment = function(domain, tissue1, tissue2, x, exons)
{
    x       = x[x$tissue1 == tissue1 & x$tissue2 == tissue2, ]
    exons   = exons[exons$tissue1 == tissue1 & exons$tissue2 == tissue2, ]
    exons11 = unique(x[x$tissue == tissue1 & x$domain == domain, "gene_id"])
    exons21 = unique(x[x$tissue == tissue2 & x$domain == domain, "gene_id"])
    exons12 = unique(exons[exons$tissue == tissue1 & !exons$gene_id %in% exons11, "gene_id"])
    exons22 = unique(exons[exons$tissue == tissue2 & !exons$gene_id %in% exons21, "gene_id"])
    
    out = data.frame(domain  = domain , 
                     tissue1 = tissue1, 
                     tissue2 = tissue2,
                     x11     = length(exons11),
                     x12     = length(exons12),
                     x21     = length(exons21),
                     x22     = length(exons22)
                    )
    
    totest = matrix(as.numeric(out[1,c("x11", "x12", "x21", "x22")]), nrow = 2, byrow = TRUE)
    test   = fisher.test(totest)
    out    = cbind(out,
                   data.frame(odds_ratio = test$estimate, 
                              log2r      = log2(test$estimate),
                              ci1        = test$conf.int[[1]],
                              ci2        = test$conf.int[[2]],
                              pval       = test$p.value
                             )
                  )
    
    
    return(out)
}

test_ipsc_cvpc_heart   = as.data.frame(rbindlist(lapply(domains2test, function(domain){calculate_enrichment(domain, "ipsc_cvpc", "heart"  , exons2domain_single, exons)})), stringsAsFactors = FALSE)
test_ipsc_cvpc_arteria = as.data.frame(rbindlist(lapply(domains2test, function(domain){calculate_enrichment(domain, "ipsc_cvpc", "arteria", exons2domain_single, exons)})), stringsAsFactors = FALSE)
test_heart_arteria     = as.data.frame(rbindlist(lapply(domains2test, function(domain){calculate_enrichment(domain, "heart"    , "arteria", exons2domain_single, exons)})), stringsAsFactors = FALSE)




test_ipsc_cvpc_heart   = test_ipsc_cvpc_heart  [ test_ipsc_cvpc_heart  $x11 >0 | test_ipsc_cvpc_heart  $x21 > 0,]
test_ipsc_cvpc_arteria = test_ipsc_cvpc_arteria[ test_ipsc_cvpc_arteria$x11 >0 | test_ipsc_cvpc_arteria$x21 > 0,]
test_heart_arteria     = test_heart_arteria    [ test_heart_arteria    $x11 >0 | test_heart_arteria    $x21 > 0,]

test_ipsc_cvpc_heart  $qval = p.adjust(test_ipsc_cvpc_heart  $pval, method = "BH")
test_ipsc_cvpc_arteria$qval = p.adjust(test_ipsc_cvpc_arteria$pval, method = "BH")
test_heart_arteria    $qval = p.adjust(test_heart_arteria    $pval, method = "BH")

nrow(test_ipsc_cvpc_heart  [test_ipsc_cvpc_heart  $qval < 0.05,])
nrow(test_ipsc_cvpc_arteria[test_ipsc_cvpc_arteria$qval < 0.05,])
nrow(test_heart_arteria    [test_heart_arteria    $qval < 0.05,])

domain_data           = unique(domains[,c("feature_name", "CDD")])
colnames(domain_data) = c("domain", "cdd")
#domain_tests          = merge(rbind(test_ipsc_cvpc_heart, test_ipsc_cvpc_arteria, test_heart_arteria), domain_data)
#domain_tests$url      = ""
#domain_tests[domain_tests$cdd != "", "url"] = paste("https://www.ncbi.nlm.nih.gov/Structure/cdd/cddsrv.cgi?uid", sub("CDD:", "", domain_tests[domain_tests$cdd != "", "cdd"]), sep = "=")

domain_tests = rbind(test_ipsc_cvpc_heart, test_ipsc_cvpc_arteria, test_heart_arteria)
fwrite(domain_tests, "pipeline/4.10.diffexp_eclip_enrichment/domain_tests.txt", sep = "\t", col.names = TRUE, row.names = FALSE)


domain_tests[ domain_tests$qval < 0.05,]

head(domain_tests)

test_ipsc_cvpc_heart[ order(test_ipsc_cvpc_heart$pval),]

exons2domain$has_domain = FALSE
exons2domain[exons2domain$domain != "", "has_domain"] = TRUE


test_exon2domain = function(tissue1, tissue2, exons2domain)
{
    x     = exons2domain[exons2domain$tissue1 == tissue1 & exons2domain$tissue2 == tissue2, ]
    out   = data.frame(x11 = nrow(x[x$tissue == tissue1 & x$has_domain == TRUE ,]),
                       x12 = nrow(x[x$tissue == tissue1 & x$has_domain == FALSE,]),
                       x21 = nrow(x[x$tissue == tissue2 & x$has_domain == TRUE ,]),
                       x22 = nrow(x[x$tissue == tissue2 & x$has_domain == FALSE,])
                      )
    
    totest = matrix(as.numeric(out[1,]), nrow = 2, byrow = TRUE)
    test   = fisher.test(totest)
    out    = cbind(out, data.frame(tissue1 = tissue1, tissue2 = tissue2, estimate = test$estimate, ci1 = test$conf.int[[1]], ci2 = test$conf.int[[2]], pval = test$p.value))
    
    return(out)
}
tests = as.data.frame(rbindlist(list(test_exon2domain("ipsc_cvpc", "heart"  , exons2domain), 
                                     test_exon2domain("ipsc_cvpc", "arteria", exons2domain), 
                                     test_exon2domain("heart"    , "arteria", exons2domain) 
                                    )), stringsAsFactors = FALSE)

tests

str(exons_unique)

str(exons2domain)

str(domains)
