==> cardiac_eqtls/6.8.examples.ipynb <==
setwd("/frazer01/projects/CARDIPS/analysis/cardiac_eqtls")

source("script/packages.R"  )
source("script/input_data.R")
source("script/functions.R" )
source("script/colors.R"    )


manifest = fread("pipeline/6.1.coloc_gwas/traits.manifest.txt"                      , sep = "\t", header = TRUE, data.table = FALSE)
fimo     = fread("pipeline/6.1.coloc_gwas//intersect_rbps/fimo2motif.txt"           , sep = "\t", header = TRUE, data.table = FALSE)
coloc    = fread("pipeline/6.1.coloc_gwas/coloc.eqtls.isoform.txt"                  , sep = "\t", header = TRUE , data.table = FALSE)
indata   = fread("pipeline/6.1.coloc_gwas//intersect_rbps/filtered_for_examples.txt", sep = "\t", header = TRUE, data.table = FALSE)
exons    = fread("/frazer01/reference/private/Gencode.v34lift37/exon_info.txt"      , sep = "\t", header = TRUE, data.table = FALSE)
toplot   = indata[ indata$pp.gwas.gene == TRUE & indata$is_cell == TRUE & indata$is_rbp == TRUE,]
toplot   = toplot[ order(toplot$coloc_gene, decreasing = TRUE),]


rbps        = data.frame(rbp = c("FMR1", "FUS", "FXR2", "HNRNPA1", "HNRNPL", "RBM5", "SRSF7"), folder = "")
rbps$folder = paste("/frazer01/home/tarthur/projects/CARDIPS/fine_mapping/analysis/2020-10-08-fgwas", tolower(rbps$rbp), "peakfiles", sep = "/")
rbps$color  = qualitative_hcl(n = nrow(rbps), palette = "Set 2")

read_rbp = function(rbp, rbps, fm)
{
    bedfiles         = list.files(rbps[rbps$rbp == rbp, "folder"], full.names = TRUE)
    indata           = as.data.frame(rbindlist(lapply(bedfiles, function(x){fread(x, sep = "\t", header = FALSE, data.table = FALSE)[,1:3]})), stringsAsFactors = FALSE)
    colnames(indata) = c("chrom", "from", "to")
    return(indata)
}

rbp_list        = lapply(rbps$rbp, function(rbp){read_rbp(rbp, rbps)})
names(rbp_list) = rbps$rbp

toplot[,c("tr2id", "gene_name", "pp", "coloc_gene", "coloc_snp")]
str(toplot)

plot_signal = function(x, id, xlims, pos, mycol, colname = "", mytext = "")
{
    x$y   = x[, mycol]
    x$cex = 1
    x$col = "#aaaaaa66"
    x[x$id == id, "cex"] = 3
    x[x$id == id, "col"] = "#ff0000ff"
    
    par(mar = c(0,6,0,1))
    plot(1,1, type = "n", xlim = xlims, ylim = c(0, max(x$y)), xlab = "", ylab = "", axes = FALSE)
    abline(h = 0)
    points(x$pos, x$y, cex = x$cex, col = x$col, pch = 16)
    
    axis(2)
    
    mtext(text = colname, side = 2, line = 2.5)
    
    text(x = xlims[[2]], y = max(x$y) * 0.5, labels = mytext, pos = 2, cex = 1.2)
    
    abline(v = pos, lty = "dashed", col = "#ff0000")
}

plot_transcript = function(transcript_id, xlims, pos, exons, rbps, rbp_list)
{
    chrom = unique(exons$chrom)
    
    par(mar = c(4,6,0,1))
    plot(1,1, type = "n", xlim = xlims, ylim = c(1, -2), xlab = "", ylab = "", axes = FALSE)
    axis(1, at = pretty(xlims, n = 5), labels = pretty(xlims, n = 5) * 1e-6)
    
    segments(x0    = min(exons$start), x1     = max(exons$end), y0      = 0.5, col = "#0000ff")
    rect    (xleft =     exons$start , xright =     exons$end , ybottom = 0.1, col = "#0000ff", ytop = 0.9, border = "#0000ff")
    
    text(x = max(exons$end), y = 0.5, labels = transcript_id, col = "#0000ff", pos = 4, cex = 1.5, font = 3)
    
    mtext(text = paste(chrom, "(Mb)"), side = 1, line = 2.5)
    
    rbps$y = -(1:nrow(rbps)) / nrow(rbps) * 2
    
    mtext(text = "Isoform", side = 2, line = 0, las = 2, at = 0.5)
    mtext(text = rbps$rbp , side = 2, line = 0, las = 2, at = rbps$y + (1/nrow(rbps)), cex= 1)
    
    for(ii in 1:nrow(rbps))
    {
        rbp   = rbps[ii, "rbp"  ]
        y     = rbps[ii, "y"    ]
        mycol = rbps[ii, "color"]
        
        this  = rbp_list[[rbp]]
        this  = this[this$chrom == chrom & this$to >= xlims[[1]] - 1e6 & this$from <= xlims[[2]] + 1e6,]
        
        rect    (xleft = this$from , xright = this$to, ybottom = y + (1/nrow(rbps) * 2), ytop = y, col = mycol, border = mycol)
    }
    abline(v = pos, lty = "dashed", col = "#ff0000")
}

plot_example = function(tr2id, toplot, coloc, manifest, exons, rbps, rbp_list)
{
    transcript_id = toplot[toplot$tr2id == tr2id, "transcript_id"]
    type          = toplot[toplot$tr2id == tr2id, "type"         ]
    gene_id       = toplot[toplot$tr2id == tr2id, "gene_id"      ]
    gene_name     = toplot[toplot$tr2id == tr2id, "gene_name"    ]
    id            = toplot[toplot$tr2id == tr2id, "id"           ]
    
    exons         = exons[exons$transcript_id == transcript_id,] 
    
    mycoloc       = coloc[coloc$transcript_id == transcript_id & coloc$type == type,]
    mycoloc       = mycoloc[order(mycoloc$PP.H4.abf, decreasing = TRUE),]
    mycoloc       = mycoloc[which.max(mycoloc$PP.H4.abf),]
    trait         = mycoloc[1, "trait"]
    trait_name    = manifest[manifest$id == trait, "description"]
    
    eqtl          = fread(paste("pipeline/3.2.eqtls/eqtls_by_gene/cardiac_eqtls.isoform" , paste("qtl", transcript_id,       "txt", sep = "."), sep = "/"), sep = "\t", header = TRUE, data.table = FALSE)
    fm            = fread(paste("pipeline/3.2.eqtls/eqtls_fine_map/cardiac_eqtls.isoform", paste("qtl", transcript_id, type, "txt", sep = "."), sep = "/"), sep = "\t", header = TRUE, data.table = FALSE)
    gwas          = fread(paste("pipeline/6.1.coloc_gwas/coloc.isoform"                  , paste("snp", transcript_id,       "txt", sep = "."), sep = "/"), sep = "\t", header = TRUE, data.table = FALSE)
    eqtl          = eqtl[eqtl$type == type,]
    eqtl$logp     = -log10(eqtl$pval)
    fm            = merge(fm, eqtl[,c("id", "pos")])
    gwas          = gwas[gwas$type == type & gwas$trait == trait,]

    xlims         = range(eqtl$pos)
    pos           = eqtl[eqtl$id == id, "pos"]
    
    totext_cell   = unlist(lapply(colnames(toplot)[grepl("^cell\\.", colnames(toplot)) == TRUE], function(cell)
    {
        out = ""
        if(toplot[toplot$tr2id == tr2id, cell] == TRUE){out = sub("^cell\\.", "", cell)}
        
        return(out)
    }))
    
    totext_rbp   = unlist(lapply(colnames(toplot)[grepl("^rbp\\.", colnames(toplot)) == TRUE], function(rbp)
    {
        out = ""
        if(toplot[toplot$tr2id == tr2id, rbp] == TRUE){out = sub("^rbp\\.", "", rbp)}
        
        return(out)
    }))
 
    png(filename = paste("pipeline/6.1.coloc_gwas/intersect_rbps/examples", paste(gene_name, gsub(":", ".", tr2id), "png", sep = "."), sep = "/"), width = 8.5, height = 8.5, units = "in", res = 300)
    
    mytext   = paste(c(gene_name, paste(totext_cell[totext_cell != ""], collapse = "\n"), paste(totext_rbp[totext_rbp != ""], collapse = "\n")), collapse = "\n")
    mylayout = c(1:4)
    layout(cbind(mylayout, mylayout + 4))

    
    plot_signal    (eqtl, id     , xlims, pos, "logp"  , "-log10 (p-value eQTL)", mytext)
    plot_signal    (fm  , id     , xlims, pos, "pp"    , "PPA (fine-map)"       , trait_name)
    plot_signal    (gwas, id     , xlims, pos, "pp_snp", "PPA (GWAS)")
    plot_transcript(transcript_id, xlims, pos, exons, rbps, rbp_list)
    
    plot_signal    (eqtl, id     , c(min(exons$start), max(exons$end)), pos, "logp"  , "-log10 (p-value eQTL)")
    plot_signal    (fm  , id     , c(min(exons$start), max(exons$end)), pos, "pp"    , "PPA (fine-map)"       )
    plot_signal    (gwas, id     , c(min(exons$start), max(exons$end)), pos, "pp_snp", "PPA (GWAS)"           )
    plot_transcript(transcript_id, c(min(exons$start), max(exons$end)), pos, exons, rbps, rbp_list)
    
    dev.off()   
}

options(repr.plot.width = 8.5, repr.plot.height = 11)


dir.create("pipeline/6.1.coloc_gwas/intersect_rbps/examples", showWarnings = FALSE)
invisible(lapply(toplot$tr2id, function(tr2id){try(plot_example(tr2id, toplot, coloc, manifest, exons, rbps, rbp_list))}))


fm[fm]

toplot[toplot$gene_name == "DDX39B",]
