==> cardiac_eqtls/7.1.trans_eqtl_rbp.ipynb <==
# Test if eQTLs for RNA binding proteins are associated with the expression of isoforms that have the RBP binding
- RBP binding: from eCLIP: `/frazer01/home/tarthur/projects/CARDIPS/rbp_analysis/sorted_bed/`


setwd("/frazer01/projects/CARDIPS/analysis/cardiac_eqtls")

source("script/packages.R"  )
source("script/input_data.R")
source("script/functions.R" )
source("script/colors.R"    )


geneinfo_gene_all   = fread("input/phenotypes/gene_info.txt"          , sep = "\t", header = TRUE , data.table = FALSE)
geneinfo_gene       = fread("pipeline/1.2.expression/gene_info.txt"   , sep = "\t", header = TRUE , data.table = FALSE)
geneinfo_isoform    = fread("pipeline/1.2.expression/isoform_info.txt", sep = "\t", header = TRUE , data.table = FALSE)
diffexp             = fread("pipeline/4.1.differential_expression/diffexp.txt", sep = "\t", header = TRUE , data.table = FALSE)
eqtl_genes          = fread("pipeline/3.2.eqtls/eqtls/cardiac_eqtls.gene.egenes.txt"         , sep = "\t", header = TRUE , data.table = FALSE)


rbp_folder         = "/frazer01/home/tarthur/projects/CARDIPS/rbp_analysis/sorted_peaks"
rbpbs_list         = list.files(rbp_folder, pattern = "bed$", full.names = TRUE)
rbp_meta           = fread("/reference/public/ENCODE/rbp_eclip_2020_09_09/metadata.tsv", sep = "\t", header = TRUE, data.table = FALSE)
rbp_meta           = rbp_meta[,c("File accession", "File format type", "File assembly", "Assay", "Biosample term name", "Experiment target")]
colnames(rbp_meta) = c("id", "filetype", "assembly", "assay", "cell_line", "gene_name")
rbp_meta$infile    = paste(rbp_folder, paste(rbp_meta$id, "bed", sep = "."), sep = "/")
rbp_meta$gene_name = gsub("-human", "", rbp_meta$gene_name)
rbp_meta           = rbp_meta[rbp_meta$infile %in% rbpbs_list & rbp_meta$assembly == "hg19",]
rbp_meta           = merge(rbp_meta, geneinfo_gene_all[geneinfo_gene_all$gene_id != "ENSG00000280987.4_7",c("gene_name", "gene_id")])
rbp_meta           = merge(rbp_meta, diffexp[diffexp$type == "gene_tpm" & diffexp$tissue1 == "ipsc_cvpc" & diffexp$tissue2 == "heart", c("gene_id", "beta", "pval", "qval", "diffexp")], all.x = TRUE)

# Intersect RBPs with isoform positions

rbp_data = as.data.frame(rbindlist(lapply(rbp_meta$infile, function(x)
{
    indata = fread(x, sep = "\t", header = FALSE, data.table = FALSE)
    indata = indata[,1:4]
    
    return(indata)
})), stringsAsFactors = FALSE)

colnames(rbp_data) = c("chrom", "from", "to", "id")
rbp_data           = rbp_data[order(rbp_data$chrom, rbp_data$from, rbp_data$to),]
rbp_bed            = paste(getwd(), "pipeline/7.1.trans/rbp/input", "rbp.bed", sep = "/")

fwrite(rbp_data, rbp_bed, sep = "\t", col.names = FALSE, row.names = FALSE)


intersected_file = paste(getwd(), "pipeline/7.1.trans/rbp/input", "intersected.bed", sep = "/")
command          = paste("bedtools", "intersect", "-loj", "-a", rbp_bed, "-b", "/reference/private/Gencode.v34lift37/isoform.bed", ">", intersected_file)

#system(command)

intersected           = fread(intersected_file, sep = "\t", header = FALSE, data.table = FALSE)
intersected           = intersected[,c(8,4)]
colnames(intersected) = c("transcript_id", "x")
intersected$rbp_name  = unlist(lapply(intersected$x, function(x){unlist(strsplit(x, "_"))[[1]]}))
#intersected           = unique(intersected)


isoform2rbp = aggregate(x ~ rbp_name + transcript_id, data = intersected[intersected$transcript_id != ".",], FUN = length)

fwrite(isoform2rbp, "pipeline/7.1.trans/rbp/input/isoform2rbp.txt", sep = "\t", col.names = TRUE, row.names = FALSE)


# Get eQTL information for each RBP

eqtl_rbp = eqtl_genes[eqtl_genes$gene_id %in% rbp_meta$gene_id, ]
fwrite(eqtl_rbp, "pipeline/7.1.trans/rbp/input/eqtl_rbp.txt", sep = "\t", col.names = TRUE, row.names = FALSE)


invisible(lapply(sort(unique(eqtl_rbp$gene_id)), function(rbp)
{
    indata = isoform2rbp[ isoform2rbp$rbp_name == geneinfo_gene[geneinfo_gene$gene_id == rbp, "gene_name"],]
    indata = indata[indata$transcript_id %in% geneinfo_isoform$transcript_id,]
    
    fwrite(indata[,c("transcript_id", "x")], paste("pipeline/7.1.trans/rbp/input/rbp2transcript", paste(rbp, "txt", sep = "."), sep = "/"), sep = "\t", col.names = TRUE, row.names = FALSE)
}))



# Run trans-analysis

run_trans_eqtls_qsub = function()
{
    
    sh_file = paste(getwd(), "pipeline/7.1.trans/script", paste("run.7.1", "rbp", "sh", sep = "."), sep = "/")
    
    writeLines(text = c("#!/usr/bin/sh",
                        "source /frazer01/home/matteo/.bashrc",
                        paste("Rscript", paste(paste(getwd(), "script", "7.2.run_trans_eqtls_rbp.R", sep = "/"),"--taskid", "$SGE_TASK_ID"))
                       ), 
               con  = sh_file, 
               sep  = "\n\n")
    
    qsub_command = paste("qsub",
                         "-l", "short",
                         "-t", paste(1, "-", nrow(eqtl_rbp), ":1", sep = ""),
                         "-tc", 500, 
                         "-o" , paste(getwd(), paste("logs/eqtls", "7.2.run_trans_eqtls_rbp", "out", sep = "."), sep = "/"),
                         "-e" , paste(getwd(), paste("logs/eqtls", "7.2.run_trans_eqtls_rbp", "err", sep = "."), sep = "/"),
                         sh_file
                        )
    
    message(qsub_command)
    system(qsub_command)
}

run_trans_eqtls_qsub() # run all genes


# Private: test trans analysis

taskid = 2

exp_all          = add_rownames(fread("pipeline/1.2.expression//use_isoform.normalized.txt", sep = "\t", header = TRUE , data.table = FALSE))


gt_folder           = paste(getwd(), "pipeline", "1.3.genotype"  , "tpm_gene"   , sep = "/")
geneinfo_gene       = fread("pipeline/1.2.expression/gene_info.txt"                          , sep = "\t", header = TRUE , data.table = FALSE)
geneinfo_isoform    = fread("pipeline/1.2.expression/isoform_info.txt"                       , sep = "\t", header = TRUE , data.table = FALSE)
eqtl_rbp            = fread("pipeline/7.1.trans/rbp/input/eqtl_rbp.txt"                      , sep = "\t", header = TRUE , data.table = FALSE)
eqtl                = eqtl_rbp[taskid, c("gene_id", "gene_name", "type", "chrom", "pos", "ref", "alt", "id", "beta", "se", "pval", "fdr", "qval", "egene")]
rbp                 = eqtl$gene_id
totest              = fread(paste("pipeline/7.1.trans/rbp/input/rbp2transcript", paste(rbp, "txt", sep = "."), sep = "/"), sep = "\t", header = TRUE , data.table = FALSE)
vars0               = readLines(paste("pipeline/3.2.eqtls", "vars", paste("cardiac_eqtls", "isoform", "txt", sep = "."), sep = "/"))
exp_rbp             = add_rownames(fread(paste(paste(getwd(), "pipeline", "1.2.expression", "tpm_gene", sep = "/"), paste(rbp, "txt", sep = "."), sep = "/"), sep = "\t", header = TRUE , data.table = FALSE))
covariates          = add_rownames(fread("pipeline/3.1.covariates/covariates.txt", sep = "\t", header = TRUE , data.table = FALSE))
metadata            =              fread("pipeline/3.1.covariates/metadata.txt"  , sep = "\t", header = TRUE , data.table = FALSE)
gtdata              = add_rownames(fread(paste(gt_folder , paste("gt_data", rbp, "txt", sep = "."), sep = "/"), sep = "\t", header = TRUE , data.table = FALSE))
covariates          =              merge(metadata, covariates, by.x = "run", by.y = "row.names")
covariates          = merge(covariates, data.frame(wgs_id = colnames(gtdata), gt = as.numeric(gtdata[eqtl$id,])))

isoform = totest[1, "transcript_id"]


calculate_trans = function(x, name, tolm, vars0)
{
    tolm$totest    = tolm[,x]
    out0           = as.data.frame(summary(lm(norm_isoform ~ ., data = tolm[,c("norm_isoform", "totest", "norm", vars0)]))$coefficients)
    out0           = out0["totest", c("Estimate", "Std. Error", "Pr(>|t|)")]
    colnames(out0) = paste(c("beta", "se", "pval"), name, sep = "_")
    
    return(out0)
}

out = as.data.frame(rbindlist(lapply(rownames(exp_all), function(isoform)
{
    expdata                  = data.frame(run = colnames(exp_all), norm_isoform = as.numeric(exp_all[isoform,]))
    tolm                     = merge(covariates, expdata, by   = "run")
    tolm                     = merge(tolm      , exp_rbp, by.x = "run", by.y = "row.names")
    out_gt                   = calculate_trans("gt"  , "gt" , tolm, vars0)
    out_ex                   = calculate_trans("norm", "exp", tolm, vars0)
    out0                     = cbind(eqtl[,c("gene_id", "gene_name", "type")], out_gt, out_ex)
    out0$transcript_id_trans = isoform  
    
    return(out0)
})), stringsAsFactors = FALSE)

fwrite(out, paste("pipeline/7.1.trans/rbp/trans_eqtls_by_rbp", paste(rbp, "txt", sep = "."), sep = "/"), sep = "\t", col.names = TRUE, row.names = FALSE)



out

metadata

str(unique(rbp_meta$gene_name))
str(unique(rbp_meta[,c("gene_name", "cell_line")]))


str(rbp_meta)
