==> cardiac_eqtls/7.3.trans_eqtl_rbp_analysis.ipynb <==
setwd("/frazer01/projects/CARDIPS/analysis/cardiac_eqtls")

source("script/packages.R"  )
source("script/input_data.R")
source("script/functions.R" )
source("script/colors.R"    )


geneinfo_gene       = fread("pipeline/1.2.expression/gene_info.txt"    , sep = "\t", header = TRUE , data.table = FALSE)
geneinfo_isoform    = fread("pipeline/1.2.expression/isoform_info.txt" , sep = "\t", header = TRUE , data.table = FALSE)
eqtl_rbp            = fread("pipeline/7.1.trans/rbp/input/eqtl_rbp.txt", sep = "\t", header = TRUE , data.table = FALSE)
gene_map            = fread("/frazer01/reference/public/cross_mapability/hg19_gencode19_75merExon_75merUTR_2mismatch_gene_mappability.txt" , sep = "\t", header = FALSE, data.table = FALSE)
gene_cross          = fread("/frazer01/reference/public/cross_mapability/hg19_gencode19_75merExon_75merUTR_2mismatch_cross_mappability.txt", sep = "\t", header = FALSE, data.table = FALSE)

colnames(gene_map  ) = c("gene_id", "mappability")
colnames(gene_cross) = c("gene_id1", "gene_id2", "cross")


eqtl_rbp$gene2type = paste(eqtl_rbp$gene_id, eqtl_rbp$type)

isoform2rbp = fread("pipeline/7.1.trans/rbp/input/isoform2rbp.txt", sep = "\t", header = TRUE , data.table = FALSE)


trans0 = as.data.frame(rbindlist(lapply(list.files("pipeline/7.1.trans/rbp/trans_eqtls_by_rbp/", full.names = TRUE), function(x)
{
    fread(x, sep = "\t", header = TRUE, data.table = FALSE)
})), stringsAsFactors = FALSE)

fwrite(trans0, "pipeline/7.1.trans/rbp/trans_eqtls.unfiltered.txt", sep = "\t", col.names = TRUE, row.names = FALSE)


# Filter by mappability, similar to GTEx methods:
- gene mappability >= 0.8
- remove all cross mapped interactions

gene_cross      $gene1 = unlist(lapply(gene_cross      $gene_id1, function(x){unlist(strsplit(x, "\\."))[[1]]}))
gene_cross      $gene2 = unlist(lapply(gene_cross      $gene_id2, function(x){unlist(strsplit(x, "\\."))[[1]]}))
gene_map        $gene  = unlist(lapply(gene_map        $gene_id , function(x){unlist(strsplit(x, "\\."))[[1]]}))
geneinfo_gene   $gene  = unlist(lapply(geneinfo_gene   $gene_id , function(x){unlist(strsplit(x, "\\."))[[1]]}))
geneinfo_isoform$gene  = unlist(lapply(geneinfo_isoform$gene_id , function(x){unlist(strsplit(x, "\\."))[[1]]}))


trans = merge(trans0, geneinfo_isoform[,c("transcript_id", "gene_id", "gene_name")], by.x = "transcript_id_trans", by.y = "transcript_id", suffixes = c("_rbp", "_trans"))

geneinfo_gene$is_map = FALSE
geneinfo_gene[geneinfo_gene$gene %in% gene_map[gene_map$mappability >= 0.8, "gene"], "is_map"] = TRUE

gene_cross$gene2gene = paste(gene_cross$gene1, gene_cross$gene2)

gene_cross_filtered = merge(gene_cross         [,c("gene1", "gene2", "cross"           )], geneinfo_gene[geneinfo_gene$is_map == TRUE, c("gene", "gene_id")], by.x = "gene1", by.y = "gene")
gene_cross_filtered = merge(gene_cross_filtered[,c("gene1", "gene2", "cross", "gene_id")], geneinfo_gene[geneinfo_gene$is_map == TRUE, c("gene", "gene_id")], by.x = "gene2", by.y = "gene", suffixes = 1:2)

gene_cross_filtered$gene2gene = paste(gene_cross_filtered$gene_id1   , gene_cross_filtered$gene_id2     )
trans              $gene2gene = paste(trans              $gene_id_rbp, trans              $gene_id_trans)

nrow(trans)
trans_filtered = trans[!trans$gene2gene %in% gene_cross_filtered$gene2gene & trans$gene_id_rbp %in% geneinfo_gene[geneinfo_gene$is_map == TRUE, "gene_id"] & trans$gene_id_trans %in% geneinfo_gene[geneinfo_gene$is_map == TRUE, "gene_id"] & trans$gene_id_rbp != trans$gene_id_trans,]
nrow(trans_filtered)


trans_filtered = merge(trans_filtered, geneinfo_gene[,c("gene_id", "chrom", "start", "end")], by.x = "gene_id_rbp"  , by.y = "gene_id")
trans_filtered = merge(trans_filtered, geneinfo_gene[,c("gene_id", "chrom", "start", "end")], by.x = "gene_id_trans", by.y = "gene_id", suffixes = c("_rbp", "_trans"))

trans_filtered = trans_filtered[trans_filtered$chrom_rbp != trans_filtered$chrom_trans,]


nrow(trans_filtered)


isoform2rbp   $rbp2iso = paste(isoform2rbp   $rbp_name     , isoform2rbp   $transcript_id      )
trans_filtered$rbp2iso = paste(trans_filtered$gene_name_rbp, trans_filtered$transcript_id_trans)
isoform2rbp   $n_bs    = isoform2rbp$x

trans2rbp = merge(trans_filtered, isoform2rbp[,c("rbp2iso", "n_bs")], all.x = TRUE)
trans2rbp[is.na(trans2rbp$n_bs) == TRUE, "n_bs"] = 0

trans2rbp$rbp2type = paste(trans2rbp$gene_id_rbp, trans2rbp$type)


str(trans2rbp)

# Find trans-eQTLs:
- RBP is eGene
- 

trans_fdr        = trans2rbp[trans2rbp$n_bs > 0 & trans2rbp$rbp2type %in% eqtl_rbp[ eqtl_rbp$egene == TRUE, "gene2type"],]
trans_fdr$fdr_gt = p.adjust(trans_fdr$pval_gt, method = "bonferroni")

nrow(trans_fdr[trans_fdr$fdr_gt < 0.05,])


head(trans_fdr[order(trans_fdr$pval_gt), c("gene_name_rbp", "gene_name_trans")])

cor.test(abs(trans2rbp$beta_gt), trans2rbp$n_bs)
t.test(abs(trans2rbp[trans2rbp$n_bs == 0, "beta_gt"]), abs(trans2rbp[trans2rbp$n_bs > 0, "beta_gt"]))

str(eqtl_rbp)
str(trans2rbp)

# FDR Correction


trans_fdr = as.data.frame(rbindlist(lapply(sort(unique(trans_filtered$rbp2type)), function(x)
{
    this = trans_filtered[trans_filtered$rbp2type == x,]
    this$fdr_gt = p.adjust(this$pval_gt, method = "BH")
    
    return(this)
})), stringsAsFactors = FALSE)

nrow(trans_fdr[trans_fdr$fdr_gt < 0.05,])


nrow(trans_filtered)

trans_filtered$fdr_gt = p.adjust(trans_filtered$pval_gt, method = "BH")

nrow(trans_filtered[trans_filtered$fdr_gt < 0.05,])

trans_filtered$fdr_gt = p.adjust(trans_filtered$pval_gt, method = "bonferroni")

nrow(trans_filtered[trans_filtered$fdr_gt < 0.05,])




a = trans_fdr[trans_fdr$fdr_gt < 0.05,]
a[order(a$pval_gt),]

str(trans)

head(trans[order(trans$pval_exp),])
