==> cardiac_eqtls/9.1.eQTLs_vs_diffexp.ipynb <==
setwd("/frazer01/projects/CARDIPS/analysis/cardiac_eqtls")

source("script/packages.R"  )
source("script/input_data.R")
source("script/functions.R" )


gene_info               = fread("pipeline/1.2.expression/gene_info.txt"    , sep = "\t", header = TRUE, data.table = FALSE)
isof_info               = fread("pipeline/1.2.expression//isoform_info.txt", sep = "\t", header = TRUE, data.table = FALSE)
gene_info$transcript_id = gene_info$gene_id

dir.create("pipeline/9.1.eqtls_vs_diffexp", showWarnings = FALSE)

# Input data

diffexp       = fread("pipeline/4.15.differential_expression_ridge/diffexp.txt"        , sep = "\t", header = TRUE , data.table = FALSE)
diffcell      = fread("pipeline/4.1.differential_expression/diffexp_cell.txt"          , sep = "\t", header = TRUE , data.table = FALSE)
eqtl_genes    = fread("pipeline/3.2.eqtls/eqtls/cardiac_eqtls.gene.egenes.txt"         , sep = "\t", header = TRUE, data.table = FALSE)
int_genes     = fread("pipeline/3.2.eqtls/eqtls/cardiac_eqtls.gene.interactions.txt"   , sep = "\t", header = TRUE, data.table = FALSE)
eqtl_isoforms = fread("pipeline/3.2.eqtls/eqtls/cardiac_eqtls.isoform.egenes.txt"      , sep = "\t", header = TRUE, data.table = FALSE)
int_isoforms  = fread("pipeline/3.2.eqtls/eqtls/cardiac_eqtls.isoform.interactions.txt", sep = "\t", header = TRUE, data.table = FALSE)
rbps          = gene_info[ gene_info$gene_name %in% readLines("pipeline/rbps.txt"), "gene_id"]

eqtl_genes   $rbp = FALSE
int_genes    $rbp = FALSE
eqtl_isoforms$rbp = FALSE
int_isoforms $rbp = FALSE

eqtl_genes   [eqtl_genes   $gene_id %in% rbps, "rbp"] = TRUE
int_genes    [int_genes    $gene_id %in% rbps, "rbp"] = TRUE
eqtl_isoforms[eqtl_isoforms$gene_id %in% rbps, "rbp"] = TRUE
int_isoforms [int_isoforms $gene_id %in% rbps, "rbp"] = TRUE


# Analysis
## Are differentially expressed genes/isoforms more likely to have stage-specific eQTLs?

test_eqtl_vs_diffexp = function(type, interaction, tissue1, tissue2, eqtls, diffexp)
{
    diffexp          = diffexp[diffexp$type == type & diffexp$tissue1 == tissue1 & diffexp$tissue2 == tissue2 & diffexp$covariate == "tissue",]
    diffexp$specific = FALSE
    
    if(tissue1 == interaction){diffexp[diffexp$diffexp == TRUE & diffexp$beta > 0, "specific"] = TRUE}
    if(tissue2 == interaction){diffexp[diffexp$diffexp == TRUE & diffexp$beta < 0, "specific"] = TRUE}
    
    eqtls_all  = unique(eqtls$transcript_id)
    eqtls_cell = unique(eqtls[eqtls$interaction == interaction & eqtls$cell == TRUE, "transcript_id"])
    genes_all  = unique(diffexp$transcript_id)
    genes_cell = unique(diffexp[diffexp$specific == TRUE, "transcript_id"])
    eqtls_all  = setdiff(eqtls_all, eqtls_cell)
    genes_all  = setdiff(genes_all, genes_cell)
    totest     = matrix(c(length(intersect(eqtls_cell, genes_cell)),
                          length(intersect(eqtls_cell, genes_all )),
                          length(intersect(eqtls_all , genes_cell)),
                          length(intersect(eqtls_all , genes_all ))
                         ), 
                        nrow = 2, byrow = TRUE)
    
    test = fisher.test(totest)
    
    return(list(totest, test))
    
}

type        = "gene_tpm"
interaction = "heart"
tissue1     = "ipsc_cvpc"
tissue2     = "heart"

test_eqtl_vs_diffexp("gene_tpm", "heart"    , "ipsc_cvpc", "heart", int_genes, diffexp)
test_eqtl_vs_diffexp("gene_tpm", "ipsc_cvpc", "ipsc_cvpc", "heart", int_genes, diffexp)


## Are cell type-specific genes/isoforms more likely to have cell type-specific eQTLs?

test_eqtl_vs_diffexp_cell = function(type, cell, eqtls, diffexp)
{
    diffexp = diffexp[diffexp$type == type & diffexp$cell_type == cell,]
    
    eqtls_all  = unique(eqtls$transcript_id)
    eqtls_cell = unique(eqtls[eqtls$interaction == cell & eqtls$cell == TRUE, "transcript_id"])
    genes_all  = unique(diffexp$transcript_id)
    genes_cell = unique(diffexp[diffexp$qval <= 0.05, "transcript_id"])
    eqtls_all  = setdiff(eqtls_all, eqtls_cell)
    genes_all  = setdiff(genes_all, genes_cell)
    totest     = matrix(c(length(intersect(eqtls_cell, genes_cell)),
                          length(intersect(eqtls_cell, genes_all )),
                          length(intersect(eqtls_all , genes_cell)),
                          length(intersect(eqtls_all , genes_all ))
                         ), 
                        nrow = 2, byrow = TRUE)
    
    test = fisher.test(totest)
    
    out = data.frame(type        = type, cell = cell,
                     eqtl1_cell1 = length(intersect(eqtls_cell, genes_cell)),
                     eqtl1_cell0 = length(intersect(eqtls_cell, genes_all )),
                     eqtl0_cell1 = length(intersect(eqtls_all , genes_cell)),
                     eqtl0_cell0 = length(intersect(eqtls_all , genes_all )),
                     or          = test$estimate,
                     ci1         = test$conf.int[[1]],
                     ci2         = test$conf.int[[2]],
                     pval        = test$p.value
                    )
    
    return(out)
    
}

tests = as.data.frame(rbindlist(lapply(sort(unique(diffcell$cell_type)), function(cell)
{
    out1 = test_eqtl_vs_diffexp_cell("gene_tpm"   , cell, int_genes   , diffcell)
    out2 = test_eqtl_vs_diffexp_cell("isoform_use", cell, int_isoforms, diffcell)
    
    return(rbind(out1, out2))
})), stringsAsFactors = FALSE)

tests      = tests[order(tests$type, tests$cell),]
tests$qval = p.adjust(tests$pval)


## Describe RBPs and their isoforms: are they more likely to have eQTLs?

table(eqtl_genes   [ eqtl_genes   $type == 0,c("egene", "rbp")])
table(eqtl_isoforms[ eqtl_isoforms$type == 0,c("egene", "rbp")])

fisher.test(table(eqtl_genes   [ eqtl_genes   $type == 0,c("egene", "rbp")]))
fisher.test(table(eqtl_isoforms[ eqtl_isoforms$type == 0,c("egene", "rbp")]))
